refactored source code into src
