+++ package.json
@@ -2,7 +2,7 @@
     "name": "cral-cli",
     "version": "1.3.0",
     "description": "A CLI for registering for courses at Rutgers Univseristy.",
-    "main": "index.js",
+    "main": "./src/index.js",
     "scripts": {
         "test": "jest --env=node",
         "start": "nodemon cral.js"
@@ -12,7 +12,7 @@
         "url": "git+https://github.com/fireteam99/RU-Reg-Helper.git"
     },
     "author": "",
-    "license": "ISC",
+    "license": "MIT",
     "bugs": {
         "url": "https://github.com/fireteam99/cral-cli/issues"
     },
@@ -31,7 +31,7 @@
         "puppeteer": "^1.13.0"
     },
     "bin": {
-        "cral": "./cral.js"
+        "cral": "./src/cral.js"
     },
     "devDependencies": {
         "eslint": "^7.17.0",
