+++ src/commands/display.js
