+++ src/commands/register.js
