+++ src/commands/reset.js
