+++ src/unused/getCASCookiesLegacy.js
