+++ src/util/codeToTerm.js
