+++ src/util/dayFromLetter.js
