+++ src/util/dayFromLetter.test.js
