+++ src/util/militaryToStandardTime.js
