+++ src/util/militaryToStandardTime.test.js
