+++ src/util/registerForIndex.js
