+++ src/util/validateIndex.js
