+++ src/util/validateIndex.test.js
